drm/bridge: cdns-mhdp: Remove dependencies to cdns-mhdp-common
Copy all necessary parts from cdns-mhdp-common and do not include
cdns-mhdp-common.h and link cdns-mhdp-common.c to cdns-mhdp any
more.
